lab 9, exercise 3 double fix.

+++ lab09/omp_apps.c
@@ -110,7 +110,7 @@ double dotp_manual_optimized(double *x, double *y, int arr_size)
     {
         int total_thread = omp_get_num_threads();
         int thread_id = omp_get_thread_num();
-        int sum = 0.0;
+        double sum = 0.0;
         for (int i = 0; i < ARRAY_SIZE; i++)
         {
             if (i % total_thread != thread_id)
